Refactor setPreDeployConfigForDotnet
so that other languages can benefit from some shared code and more easily follow the same pattern
